+++ src/gui/settings.py
@@ -21,6 +21,7 @@ class SParamViewerAppSettings(AppSettings):
     ext_editor_cmd: str = ''
     plot_style: str = None
     plot_mark_points: bool = False
+    last_directories: list[str] = []
 
 
 Settings = SParamViewerAppSettings(